@@ -4,7 +4,8 @@ version: 1.0.5
 homepage: https://github.com/psovit/rich_code_editor
 
 environment:
-  sdk: ">=2.2.2 <3.0.0"
+#  sdk: ">=2.2.2 <3.0.0"
+  sdk: ">=2.12.0 <4.0.0"
 
 dependencies:
   flutter:
